# core-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Freeze:
    properties:
      canceled_at:
        type: string
      created_at:
        type: string
      ends_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      starts_at:
        type: string
      user_id:
        type: integer
    type: object
  domain.Member:
    properties:
      created_at:
        type: string
      effective_goal:
        type: integer
      first_name:
        type: string
      freeze:
        $ref: '#/definitions/domain.Freeze'
      freeze_cooldown_until:
        type: string
      goal_per_period:
        type: integer
      has_avatar:
        type: boolean
      has_closed_period:
        type: boolean
      id:
        type: integer
      joined_at:
        type: string
      last_closed_period_failed:
        type: boolean
      period_ends_at:
        type: string
      photo_url:
        type: string
      rank:
        type: string
      sport_emoji:
        type: string
      sport_name:
        type: string
      status_emoji:
        type: string
      status_text:
        type: string
      streak:
        type: integer
      theme:
        type: string
      username:
        type: string
      workouts_count:
        type: integer
    type: object
  domain.Room:
    properties:
      created_at:
        type: string
      creator_id:
        type: integer
      goal_per_period:
        type: integer
      has_avatar:
        type: boolean
      id:
        type: integer
      invite_code:
        type: string
      kind:
        type: string
      name:
        type: string
      period_days:
        type: integer
      votes_required:
        type: integer
    type: object
...
  domain.RoomWithProgress:
    properties:
      created_at:
        type: string
      creator_id:
        type: integer
      goal_per_period:
        type: integer
      has_avatar:
        type: boolean
      id:
        type: integer
      invite_code:
        type: string
      kind:
        type: string
      members_count:
        type: integer
      my_goal:
        type: integer
      name:
        type: string
      period_days:
        type: integer
      period_ends_at:
        type: string
      streak:
        type: integer
      votes_required:
        type: integer
      workouts_count:
        type: integer
    type: object
...
  httpapi.CreateRoomRequest:
    properties:
      goal_per_period:
        example: 3
        maximum: 100
        minimum: 1
        type: integer
      kind:
        enum:
        - open
        - invite
        example: open
        type: string
      name:
        example: Iron buddies
        type: string
      period_days:
        example: 7
        maximum: 90
        minimum: 1
        type: integer
      votes_required:
        example: 2
        maximum: 20
        minimum: 1
        type: integer
    type: object
  httpapi.ErrorResponse:
    properties:
      error:
        example: room not found
        type: string
    type: object
...
  httpapi.JoinByCodeRequest:
    properties:
      invite_code:
        example: 7HKPQ2XW
        type: string
    type: object
...
  httpapi.RoomDetailResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/domain.Member'
        type: array
      room:
        $ref: '#/definitions/domain.Room'
    type: object
...
  httpapi.UpdateMembershipRequest:
    properties:
      goal_per_period:
        example: 2
        maximum: 100
        minimum: 1
        type: integer
      sport_emoji:
        example: "\U0001F9D7"
        type: string
      sport_name:
        example: climbing
        type: string
    type: object
  httpapi.UpdateRoomRequest:
    properties:
      goal_per_period:
        example: 3
        maximum: 100
        minimum: 1
        type: integer
      kind:
        enum:
        - open
        - invite
        example: open
        type: string
      name:
        example: Iron buddies
        type: string
      period_days:
        example: 7
        maximum: 90
        minimum: 1
        type: integer
      votes_required:
        example: 2
        maximum: 20
        minimum: 1
        type: integer
    type: object
...
info:
  contact: {}
  description: 'Core service of BuddyGym: auth, users, rooms, rewards, checkin proxy.'
  title: BuddyGym Core API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: Rooms the user belongs to, with the current period workout counter.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.RoomWithProgress'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Creates a room and enrolls the creator. Returns the room with its
        invite code.
      parameters:
      - description: room settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a room
      tags:
      - rooms
  /rooms/{id}:
    delete:
      description: Deletes a room and its memberships. Only the creator may do this.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a room
      tags:
      - rooms
    get:
      description: Room with members. Invite-only rooms are visible to members only;
        the invite code is hidden from non-members.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.RoomDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get room details
      tags:
      - rooms
    patch:
      consumes:
      - application/json
      description: Updates room settings. Only the creator may do this.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: room settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a room
      tags:
      - rooms
...
  /rooms/{id}/join:
    post:
      description: Joins an open room by id. Invite-only rooms require POST /rooms/join
        with a code. Idempotent.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join an open room
      tags:
      - rooms
  /rooms/{id}/leave:
    post:
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Leave a room
      tags:
      - rooms
  /rooms/{id}/membership:
    patch:
      consumes:
      - application/json
      description: Sets the member's personal sport and workout goal for this room.
        Null goal falls back to the room goal, empty strings clear the sport.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: personal settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.UpdateMembershipRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update my settings in a room
      tags:
      - rooms
  /rooms/join:
    post:
      consumes:
      - application/json
      description: Joins any room using its invite code. Idempotent.
      parameters:
      - description: invite code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httpapi.JoinByCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join a room by invite code
      tags:
      - rooms
  /rooms/open:
    get:
      description: Rooms anyone can join, excluding the ones the user is already in.
        Invite codes are not returned.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Room'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List open rooms
      tags:
      - rooms
...
securityDefinitions:
  BearerAuth:
    description: 'JWT from POST /auth/telegram: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"